refactor: updated directory stucture, code cleanup
