+++ frontend/src/app/dashboard/layout.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import Navbar from "@/components/Navbar";
 
 export default function DashboardLayout({
