+++ frontend/src/app/page.tsx
@@ -1,6 +1,5 @@
 import { redirect } from "next/navigation";
 
-//TODO: handle gracefully redirecting from /* to /dashboard/{default}
 export default function Home() {
   redirect("/dashboard/pharmacy");
 }
